backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
# python-dotenv==1.0.0
openai==1.3.5
#google-generativeai==0.3.1
azure-ai-textanalytics==5.3.0
azure-identity==1.15.0
psycopg2-binary==2.9.9
pymongo==4.6.0
pymilvus>=2.6.0
sqlalchemy==2.0.23
motor==3.3.2
httpx==0.25.2
python-multipart==0.0.6
# grpcio will be installed automatically by pymilvus with Python 3.12 compatible wheels
